tests on Groups in their own file
